Truncate text until labels fit within states
When the timeline chart is zoom in/out, the label of each state
will be modified to fit the width of the state. When the width
of the state is smaller than the width of the label, the text is
truncated. To determine how much text is truncated, the timeline
chart uses the ratio of width of label text / width of state.
However, since the width of each charater is different, this ratio
does not well-represent how much text can be displayed in the label,
causing text to overflow out of the state.

This commit tries to resolve the issue by truncating the label
one letter at a time (after applying the ratio) until the label text
fits within its state.

Signed-off-by: Hoang Thuan Pham <hoang.pham@calian.ca>