Rescale timeline-chart post zoom using PIXI.Container scale property
This commit uses PIXI.Container scale property to rescale the chart after
each zoom in/zoom out to instead of the using the manually repositioning
each elements in the chart to improve the performance of the timeline chart.
The scale of the time-graph-chart is changed everytime the view range
changes, and is reset once the server has replied to the front-end.

This commit removes the method updateScaleAndPosition(), which was used
to manually rescaling the chart.

Signed-off-by: Hoang Thuan Pham <hoang.pham@calian.ca>